dvbevrec3: fix recording failure on EIT updates
When we start waiting for a new program's start (for scheduled recordings),
a version of EIT present/folloing sections can be received in either order,
so it is possible that we receive two succesive "following" sections
(the old, and the updated ones) without receiving any "present" section.
This had caused empty recordings due to the wrong processing of
EIT updates.